# Mimir query engine: improve performance of binary operations