#### What this PR does

This PR applies the same technique as https://github.com/grafana/mimir/pull/8671 for binary operations.

This improves performance by up to 7% in our benchmarks, particularly for cases where not all series from a side are used. Peak memory utilisation is unchanged.

<details>
<summary>Benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
                                                                                                                  │   main.txt    │             after.txt             │
                                                                                                                  │    sec/op     │   sec/op     vs base              │
Query/a_1_-_b_1,_instant_query/Mimir-10                                                                              287.4µ ±  2%   305.4µ ± 2%  +6.25% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps/Mimir-10                                                                 307.5µ ±  2%   309.9µ ± 1%       ~ (p=0.180 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps/Mimir-10                                                                456.7µ ±  1%   455.9µ ± 1%       ~ (p=1.000 n=6)
Query/a_100_-_b_100,_instant_query/Mimir-10                                                                          1.647m ±  1%   1.646m ± 1%       ~ (p=0.818 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps/Mimir-10                                                             2.865m ±  1%   2.862m ± 2%       ~ (p=0.818 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps/Mimir-10                                                            13.64m ±  1%   13.74m ± 1%       ~ (p=0.093 n=6)
Query/a_2000_-_b_2000,_instant_query/Mimir-10                                                                        24.69m ±  1%   24.33m ± 1%  -1.48% (p=0.015 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps/Mimir-10                                                           47.43m ±  1%   46.91m ± 0%  -1.09% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps/Mimir-10                                                          250.3m ±  0%   250.4m ± 2%       ~ (p=0.699 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps/Mimir-10                                                               1.835m ±  1%   1.823m ± 2%       ~ (p=0.132 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps/Mimir-10                                                           118.7m ±  1%   119.0m ± 1%       ~ (p=0.180 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps/Mimir-10                                                          2.332 ± 10%    2.333 ± 1%       ~ (p=0.818 n=6)
Query/nh_1_+_nh_1,_instant_query/Mimir-10                                                                            290.6µ ±  6%   291.3µ ± 2%       ~ (p=0.818 n=6)
Query/nh_1_+_nh_1,_range_query_with_100_steps/Mimir-10                                                               435.2µ ±  1%   427.9µ ± 1%  -1.68% (p=0.004 n=6)
Query/nh_1_+_nh_1,_range_query_with_1000_steps/Mimir-10                                                              1.538m ±  1%   1.535m ± 2%       ~ (p=0.699 n=6)
Query/nh_100_+_nh_100,_instant_query/Mimir-10                                                                        3.835m ±  1%   3.797m ± 0%  -0.99% (p=0.004 n=6)
Query/nh_100_+_nh_100,_range_query_with_100_steps/Mimir-10                                                           15.01m ±  0%   14.83m ± 1%  -1.21% (p=0.002 n=6)
Query/nh_100_+_nh_100,_range_query_with_1000_steps/Mimir-10                                                          107.0m ±  1%   106.4m ± 1%       ~ (p=0.394 n=6)
Query/nh_2000_+_nh_2000,_instant_query/Mimir-10                                                                      67.07m ±  1%   66.88m ± 1%       ~ (p=0.699 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_100_steps/Mimir-10                                                         269.9m ±  1%   270.8m ± 1%       ~ (p=1.000 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_1000_steps/Mimir-10                                                         2.008 ±  1%    2.002 ± 2%       ~ (p=0.394 n=6)
Query/nh_1_/_a_1,_instant_query/Mimir-10                                                                             273.3µ ±  1%   273.4µ ± 1%       ~ (p=0.937 n=6)
Query/nh_1_/_a_1,_range_query_with_100_steps/Mimir-10                                                                356.9µ ±  3%   346.1µ ± 1%  -3.01% (p=0.002 n=6)
Query/nh_1_/_a_1,_range_query_with_1000_steps/Mimir-10                                                               1.050m ±  1%   1.035m ± 1%  -1.42% (p=0.002 n=6)
Query/nh_100_/_a_100,_instant_query/Mimir-10                                                                         2.798m ±  6%   2.741m ± 0%  -2.04% (p=0.002 n=6)
Query/nh_100_/_a_100,_range_query_with_100_steps/Mimir-10                                                            9.367m ±  7%   9.115m ± 1%  -2.70% (p=0.002 n=6)
Query/nh_100_/_a_100,_range_query_with_1000_steps/Mimir-10                                                           64.06m ±  4%   60.03m ± 2%  -6.29% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_instant_query/Mimir-10                                                                       47.05m ±  1%   45.92m ± 0%  -2.41% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_range_query_with_100_steps/Mimir-10                                                          165.1m ±  1%   159.3m ± 1%  -3.51% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_range_query_with_1000_steps/Mimir-10                                                          1.208 ±  2%    1.169 ± 1%  -3.19% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query/Mimir-10                                                           1.632m ±  1%   1.600m ± 1%  -2.00% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps/Mimir-10                                              2.713m ±  1%   2.654m ± 1%  -2.19% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps/Mimir-10                                             12.38m ±  0%   12.06m ± 1%  -2.58% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query/Mimir-10                                                             16.24m ±  1%   15.52m ± 1%  -4.42% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps/Mimir-10                                                29.79m ±  6%   29.09m ± 1%  -2.35% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps/Mimir-10                                               153.4m ±  4%   150.8m ± 1%  -1.72% (p=0.015 n=6)
Query/a_2000{l="1234"}_-_b_2000,_instant_query/Mimir-10                                                              6.585m ±  4%   6.123m ± 1%  -7.01% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_100_steps/Mimir-10                                                 9.795m ±  9%   9.580m ± 1%  -2.20% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_1000_steps/Mimir-10                                                46.86m ±  1%   46.51m ± 2%  -0.75% (p=0.041 n=6)
Query/a_2000_-_b_2000{l="1234"},_instant_query/Mimir-10                                                              6.324m ±  1%   6.068m ± 1%  -4.04% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_100_steps/Mimir-10                                                 9.671m ±  1%   9.400m ± 1%  -2.80% (p=0.002 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_1000_steps/Mimir-10                                                46.05m ±  1%   45.68m ± 1%       ~ (p=0.093 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_instant_query/Mimir-10                                                          242.7µ ±  1%   247.2µ ± 3%  +1.86% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_100_steps/Mimir-10                                             268.7µ ±  1%   274.5µ ± 1%  +2.14% (p=0.002 n=6)
Query/rate(a_1[1m])_+_rate(b_1[1m]),_range_query_with_1000_steps/Mimir-10                                            432.9µ ±  9%   438.3µ ± 1%       ~ (p=0.240 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_instant_query/Mimir-10                                                      865.1µ ±  4%   852.7µ ± 5%  -1.42% (p=0.041 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_100_steps/Mimir-10                                         2.346m ±  0%   2.376m ± 1%  +1.25% (p=0.015 n=6)
Query/rate(a_100[1m])_+_rate(b_100[1m]),_range_query_with_1000_steps/Mimir-10                                        15.90m ±  6%   15.48m ± 2%       ~ (p=0.394 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_instant_query/Mimir-10                                                   10.096m ±  2%   9.663m ± 4%  -4.29% (p=0.015 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_100_steps/Mimir-10                                       39.30m ±  6%   38.95m ± 0%       ~ (p=0.065 n=6)
Query/rate(a_2000[1m])_+_rate(b_2000[1m]),_range_query_with_1000_steps/Mimir-10                                      291.4m ±  2%   289.9m ± 2%       ~ (p=0.699 n=6)
Query/sum(a_1_+_b_1),_instant_query/Mimir-10                                                                         267.2µ ±  2%   279.0µ ± 4%  +4.41% (p=0.009 n=6)
Query/sum(a_1_+_b_1),_range_query_with_100_steps/Mimir-10                                                            294.8µ ±  3%   292.8µ ± 2%       ~ (p=0.818 n=6)
Query/sum(a_1_+_b_1),_range_query_with_1000_steps/Mimir-10                                                           454.4µ ± 45%   424.4µ ± 2%  -6.62% (p=0.004 n=6)
Query/sum(a_100_+_b_100),_instant_query/Mimir-10                                                                     1.640m ±  1%   1.620m ± 1%  -1.23% (p=0.009 n=6)
Query/sum(a_100_+_b_100),_range_query_with_100_steps/Mimir-10                                                        2.883m ±  1%   2.860m ± 1%  -0.80% (p=0.041 n=6)
Query/sum(a_100_+_b_100),_range_query_with_1000_steps/Mimir-10                                                       14.08m ±  3%   13.95m ± 2%       ~ (p=0.065 n=6)
Query/sum(a_2000_+_b_2000),_instant_query/Mimir-10                                                                   25.16m ±  4%   24.87m ± 9%       ~ (p=0.093 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_100_steps/Mimir-10                                                      48.10m ±  2%   47.89m ± 9%       ~ (p=0.818 n=6)
Query/sum(a_2000_+_b_2000),_range_query_with_1000_steps/Mimir-10                                                     260.7m ±  3%   260.0m ± 1%       ~ (p=0.485 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_instant_query/Mimir-10                        256.6µ ±  3%   251.7µ ± 1%       ~ (p=0.132 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_range_query_with_100_steps/Mimir-10           286.2µ ±  3%   279.5µ ± 1%       ~ (p=0.180 n=6)
Query/sum_without_(l)(rate(a_1[1m]))_/_sum_without_(l)(rate(b_1[1m])),_range_query_with_1000_steps/Mimir-10          451.5µ ±  2%   451.9µ ± 1%       ~ (p=0.818 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_instant_query/Mimir-10                    803.3µ ±  5%   801.2µ ± 5%  -0.27% (p=0.041 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_range_query_with_100_steps/Mimir-10       2.186m ±  0%   2.184m ± 2%       ~ (p=0.818 n=6)
Query/sum_without_(l)(rate(a_100[1m]))_/_sum_without_(l)(rate(b_100[1m])),_range_query_with_1000_steps/Mimir-10      14.36m ±  0%   14.30m ± 1%       ~ (p=0.310 n=6)
Query/sum_without_(l)(rate(a_2000[1m]))_/_sum_without_(l)(rate(b_2000[1m])),_instant_query/Mimir-10                  8.398m ± 13%   8.372m ± 1%       ~ (p=0.132 n=6)
Query/sum_without_(l)(rate(a_2000[1m]))_/_sum_without_(l)(rate(b_2000[1m])),_range_query_with_100_steps/Mimir-10     38.25m ±  5%   36.01m ± 0%  -5.86% (p=0.002 n=6)
Query/sum_without_(l)(rate(a_2000[1m]))_/_sum_without_(l)(rate(b_2000[1m])),_range_query_with_1000_steps/Mimir-10    282.9m ±  3%   276.0m ± 1%  -2.43% (p=0.002 n=6)
geomean                                                                                                              7.594m         7.502m       -1.21%

                                                                                                                  │    main.txt    │              after.txt               │
                                                                                                                  │      B/op      │     B/op       vs base               │
Query/a_1_-_b_1,_instant_query/Mimir-10                                                                              42.51Ki ±  0%   43.50Ki ±  0%   +2.34% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_100_steps/Mimir-10                                                                 45.65Ki ±  0%   46.65Ki ±  0%   +2.18% (p=0.002 n=6)
Query/a_1_-_b_1,_range_query_with_1000_steps/Mimir-10                                                                69.34Ki ±  0%   70.32Ki ±  0%   +1.41% (p=0.002 n=6)
Query/a_100_-_b_100,_instant_query/Mimir-10                                                                          345.0Ki ±  0%   340.8Ki ±  0%   -1.23% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_100_steps/Mimir-10                                                             660.7Ki ±  0%   656.1Ki ±  0%   -0.69% (p=0.002 n=6)
Query/a_100_-_b_100,_range_query_with_1000_steps/Mimir-10                                                            2.900Mi ±  0%   2.895Mi ±  0%   -0.17% (p=0.041 n=6)
Query/a_2000_-_b_2000,_instant_query/Mimir-10                                                                        6.059Mi ±  0%   5.938Mi ±  0%   -2.00% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_100_steps/Mimir-10                                                           12.11Mi ±  0%   11.97Mi ±  1%   -1.16% (p=0.002 n=6)
Query/a_2000_-_b_2000,_range_query_with_1000_steps/Mimir-10                                                          57.05Mi ±  0%   56.91Mi ±  1%        ~ (p=0.485 n=6)
Query/a_1_-_b_1,_range_query_with_10000_steps/Mimir-10                                                               443.6Ki ±  1%   444.9Ki ±  0%        ~ (p=0.394 n=6)
Query/a_100_-_b_100,_range_query_with_10000_steps/Mimir-10                                                           36.04Mi ±  0%   36.09Mi ±  0%        ~ (p=0.065 n=6)
Query/a_2000_-_b_2000,_range_query_with_10000_steps/Mimir-10                                                         1.191Gi ±  0%   1.192Gi ±  0%        ~ (p=0.485 n=6)
Query/nh_1_+_nh_1,_instant_query/Mimir-10                                                                            74.29Ki ±  0%   75.29Ki ±  0%   +1.35% (p=0.002 n=6)
Query/nh_1_+_nh_1,_range_query_with_100_steps/Mimir-10                                                               211.1Ki ±  0%   212.1Ki ±  0%   +0.48% (p=0.002 n=6)
Query/nh_1_+_nh_1,_range_query_with_1000_steps/Mimir-10                                                              1.343Mi ±  0%   1.344Mi ±  0%   +0.05% (p=0.002 n=6)
Query/nh_100_+_nh_100,_instant_query/Mimir-10                                                                        1.955Mi ±  0%   1.950Mi ±  0%   -0.26% (p=0.002 n=6)
Query/nh_100_+_nh_100,_range_query_with_100_steps/Mimir-10                                                           14.58Mi ±  0%   14.57Mi ±  0%   -0.04% (p=0.004 n=6)
Query/nh_100_+_nh_100,_range_query_with_1000_steps/Mimir-10                                                          127.7Mi ±  0%   127.7Mi ±  0%        ~ (p=0.818 n=6)
Query/nh_2000_+_nh_2000,_instant_query/Mimir-10                                                                      38.13Mi ±  1%   38.07Mi ±  1%        ~ (p=0.240 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_100_steps/Mimir-10                                                         291.9Mi ±  0%   291.5Mi ±  0%        ~ (p=0.132 n=6)
Query/nh_2000_+_nh_2000,_range_query_with_1000_steps/Mimir-10                                                        2.515Gi ±  0%   2.515Gi ±  0%        ~ (p=1.000 n=6)
Query/nh_1_/_a_1,_instant_query/Mimir-10                                                                             58.52Ki ±  0%   59.52Ki ±  0%   +1.71% (p=0.002 n=6)
Query/nh_1_/_a_1,_range_query_with_100_steps/Mimir-10                                                                139.1Ki ±  0%   140.1Ki ±  0%   +0.71% (p=0.002 n=6)
Query/nh_1_/_a_1,_range_query_with_1000_steps/Mimir-10                                                               833.8Ki ±  0%   834.5Ki ±  0%   +0.08% (p=0.009 n=6)
Query/nh_100_/_a_100,_instant_query/Mimir-10                                                                         1.159Mi ±  0%   1.155Mi ±  0%   -0.42% (p=0.002 n=6)
Query/nh_100_/_a_100,_range_query_with_100_steps/Mimir-10                                                            8.652Mi ±  0%   8.646Mi ±  0%   -0.06% (p=0.002 n=6)
Query/nh_100_/_a_100,_range_query_with_1000_steps/Mimir-10                                                           76.48Mi ±  0%   76.46Mi ±  0%        ~ (p=0.937 n=6)
Query/nh_2000_/_a_2000,_instant_query/Mimir-10                                                                       22.37Mi ±  1%   22.25Mi ±  0%   -0.55% (p=0.002 n=6)
Query/nh_2000_/_a_2000,_range_query_with_100_steps/Mimir-10                                                          174.1Mi ±  0%   174.0Mi ±  0%        ~ (p=0.394 n=6)
Query/nh_2000_/_a_2000,_range_query_with_1000_steps/Mimir-10                                                         1.510Gi ±  0%   1.510Gi ±  0%        ~ (p=0.310 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_instant_query/Mimir-10                                                           264.3Ki ±  0%   260.4Ki ±  0%   -1.45% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_100_steps/Mimir-10                                              446.9Ki ±  0%   443.4Ki ±  0%   -0.76% (p=0.002 n=6)
Query/a_100{l=~"[13579]."}_-_b_100,_range_query_with_1000_steps/Mimir-10                                             1.788Mi ±  0%   1.786Mi ±  0%        ~ (p=0.240 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_instant_query/Mimir-10                                                             4.084Mi ±  0%   3.978Mi ±  0%   -2.60% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_100_steps/Mimir-10                                                7.471Mi ±  0%   7.366Mi ±  0%   -1.41% (p=0.002 n=6)
Query/a_2000{l=~"1..."}_-_b_2000,_range_query_with_1000_steps/Mimir-10                                               32.69Mi ±  0%   32.61Mi ±  0%   -0.25% (p=0.002 n=6)
Query/a_2000{l="1234"}_-_b_2000,_instant_query/Mimir-10                                                              1.799Mi ± 17%   1.744Mi ± 11%        ~ (p=0.240 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_100_steps/Mimir-10                                                 2.296Mi ± 22%   2.141Mi ± 11%   -6.76% (p=0.026 n=6)
Query/a_2000{l="1234"}_-_b_2000,_range_query_with_1000_steps/Mimir-10                                                6.653Mi ±  1%   6.598Mi ±  1%   -0.83% (p=0.041 n=6)
Query/a_2000_-_b_2000{l="1234"},_instant_query/Mimir-10                                                              2.048Mi ±  7%   1.839Mi ±  8%  -10.22% (p=0.009 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_100_steps/Mimir-10                                                 2.220Mi ± 26%   2.269Mi ± 11%        ~ (p=0.394 n=6)
Query/a_2000_-_b_2000{l="1234"},_range_query_with_1000_steps/Mimir-10                                                6.633Mi ±  1%   6.562Mi ±  1%   -1.07% (p=0.002 n=6)
…